fix : separate function files by function 'statement.js', 'createStatementData.js'
