+++ ch_01/ex/js/statement.js
@@ -1,0 +1,33 @@
+import createStatementData from "./createStatementData.js"
+
+function statement(invoice, plays){
+    return renderPlainText(createStatementData(invoice, plays));
+    function renderPlainText(data, plays){
+        let result = `청구 내역 (고객명 : ${data.customer})\n`;
+        for(let perf of data.performances){
+            result += ` ${perf.play.name}: ${usd(perf.amount)} (${perf.audience}석)\n`;
+        }
+        result += `총액: ${usd(data.totalAmount)}\n`;
+        result += `적립 포인트 : ${data.totalVolumeCredits}점\n`;
+        return result;
+    }
+    function usd(aNumber) {
+        return new Intl.NumberFormat("en-US",
+            {
+                style: "currency", currency: "USD",
+                minimumFractionDigits: 2
+            }).format(aNumber / 100);
+    }
+}
+
+function readJSON(file, callback) {
+    let rawFile = new XMLHttpRequest();
+    rawFile.overrideMimeType("application/json");
+    rawFile.open("GET", file, true);
+    rawFile.onreadystatechange = function() {
+        if (rawFile.readyState === 4 && rawFile.status === 200) {
+            callback(rawFile.responseText);
+        }
+    }
+    rawFile.send(null);
+}
