+++ ch_01/ex/pages/index.html
@@ -8,7 +8,7 @@
 
 
     <button onclick="test();">test</button>
-    <script type="text/javascript" src="../js/app.js"></script>
+    <script type="text/javascript" src="../js/statement.js"></script>
     <script type="text/javascript">
 
         function test() {
